Implement remove_and_add() method in ScoringList
This function allows caller to delete selected scoring entry and replace
it with another one.  The selection criterion is key/value pair in 'data'
attribute.  The idea is that the caller will build ScoringList while
tagging the entries with key/value pairs that are unique and later used
for this remove+add operation.

This was written for ClanAscension tracker to replace the scoring
ascensions by newer, higher scoring ascensions.