Sync achievements in config with ones in game, add bitmasks
A few duplicate achievements were in here; remove them. In most places
where the order of this disagreed with the order of the achievement
constants in the game, the game's constants were reordered; but in
places where it made sense to group similar achievements together, the
order in the game was preserved and it was changed here.

Several achievements are as yet unimplemented in the game, but will be
soon and if I know exactly where they're going in the order of
achievement constants, I added in the correct bitmask here. Some of
these, like killing specific uniques, I don't know the exact placement
in the game yet so I didn't assign bitmasks.

There were also a few achievements implemented in game that weren't in
here for some reason, so I added those in.