Solve the problem with GreatFoo reporting wrong achievement date
The code that was creating the scoring entries was a closure created when
the _first_ game was seen on the input and therefore by referencing this
game the scoring entry got the endtime of that first game.  But we need
to have last game's endtime as correct 'when' field in the scoring entry,
so we now allow passing of an argument from MultiSet's track() method to
the callback and use this to pass current game.