Fix tracking of the All Conducts trophy in AllCats tracker
We were requiring the 'zen' and 'nude' conducts, which is probably
undesirable (too hard).  Therefore we define conducts.all in the
configuration that defines what constitutes 'all conducts' for the
purpose of the trophy.  This needs to set the relevant MultiSet tracker
to 'loose' mode in case these conducts appeared anyway.