Page content update: add a header/navigation bar
This changes nothing in the actual page contents except to insert the
header at the start of the body. I wasn't able to make it work for the
individual player and individual clan templates for some reason.

Also adds in the SVG logo graphic that the header incorporates,
featuring an @ encircled by a laurel wreath.

This commit deliberately changes NO CSS.