Add javascript countdown timer to front page
The javascript is pretty much copied from the devnull tribute
tournament. It works, don't touch.

Text is changed a bit though, it hardcodes the start and end date
strings because I found the existing use of providing the start time in
your own local time without specifying a timezone to be confusing.