Create shell script that converts achievements from TNNT game into JSON
Since the game itself will now be storing achievement names and
descriptions, the idea is to have the game be the single point of truth
for that data, and avoid having to synchronize them with config.json
over here every time they are rearranged or new ones added. This script
reaches into decl.c where the achievements are stored, generates the
proper tnntachieveX and bitmask for it (that's the important part), and
wraps it all up as JSON for the scoreboard to use.