Add prototype conduct Z-scoring algorithm
Attempt to implement the algorithm described by @copperwater.
Every time a conduct is repeated, its value decreases in the series
1, 1/2, 1/3, ...
However, the algorithm attempts to find the optimal sequence in which
to count games in order to deliver the highest possible score with
a given set of ascensions, depending on the conducts kept.
I've not completely understood yet why the order matters, but it may
be something to do with the conducts not all having equal value -
otoh, individual conducts and their z-factors don't cross-talk & the
algorithm design of picking the highest value (most conducts) of the
remaining games each round will likely maximise the rate of decay.

Having said all that, @copperwater is the one who's really looked at
this and tried to make mathematical proofs and so on, so probably knows
what he's doing. :)

Signed-off-by: Joanna Doyle <jjadoyle@gmail.com>