Switch to simpler Greedy Conduct Z-Score impl
With the previous algorithm it looked as if we weren't actually saving
the grid for anything useful. Also, nd-arrays are a bit clumsy to work
with in perl. Instead, switch to a hash based method where each conduct
has an associated Z-factor and multiplier key/value pair.
The simplified algorithm is also shorter and a bit more readable.
Remove some hard-coded stuff from config.json-example that is no longer
necessary.

Signed-off-by: Joanna Doyle <jjadoyle@gmail.com>