Change lookup keys for previous ascensions
ScoringList::remove_and_add() and get_score_by_key() were unable
to find games saved with the actual $game ref as a lookup key value.
This is possibly because of equality overloading of blessed refs.
Instead, switch to keys of the form clan/playername-index, where index
is the index of the ascension in the clan or player ascension list.

Signed-off-by: Joanna Doyle <jjadoyle@gmail.com>