Fix 0-point conduct trophies on player page
I had been using the $player->add_score() feature to add and
retroactively update game score multipliers for conduct Z-scoring. These
were the entries appearing on the player page. In fact this was not
necessary, because I alrady had a reference for each game saved in the
tracked cond_ascs list.
Thus, instead, perform updates with
$cond_asc[$i]->{'game'}->remove_and_add() - these may not be used again
but it seems untidy not to update them.

Signed-off-by: Joanna Doyle <jjadoyle@gmail.com>