Fix: add data export code for new Mercy trophies
New mercy trophies could not be shown on the trophies webpage, because the
relevant data structures weren't exported in a way that was accessible
to templates/trophies.tt, which expects a trophies data structure
containing a players element and a clans element. Each of these contains
a list of trophies by name, these in turn contain player name or clan id
lists. E.g. trophies.players.mercy.wiz = ['aoei', 'Luxidream', 'stenno']
would be an expected data field.

In the course of working on this, I noticed some hard-coded trophy
export code in TNNT/Player.pm and TNNT/ClanList.pm, this has been
updated to include the new trophies, but would ideally be reworked in
order to be configurable by the config.json instead. To-do in December.
For now, add the new trophies to those hard-coded. This doesn't help the
trophies page but may be important for exported JSON or for other pages.

The actual data required for the trophies page is exported by
TNNT/Score.pm, with calls to the individual trackers for obtaining
relevant data, e.g. through a $tracker->export() subroutine.
Add an export subroutine to TNNT/Tracker/Mercy.pm, and related player
and clan tracking functionality to facilitate this (very similar to what
TNNT/Tracker/GreatFoo.pm does).

Lastly, change TNNT/Tracker/Merciful.pm to TNNT/Tracker/Mercy.pm, and
some of the other references to "merciful", to make the name consistent
with the trophy title in the JSON and on the trophies page.
Testing shows the desired data structures should be available to
templates/trophies.tt, but additional code needs to be modified
(@copperwater is on this) to actually display them on the page.

Reported-by: @copperwater
Signed-off-by: @aoeixsz4 / Joanna Doyle <jjadoyle@gmail.com>