Improve the readability of the Trophies page
Trophies that can have multiple players now generate with a spaced middot
between each pair of players. I've noticed that sometimes this places a
middot at the beginning or end of a line, but I don't know of a way to
intelligently apply them only at non-line-breaks.

Trophies that can have multiple clans force the clans shown to display
on separate lines.

Both of the above categories now use a font weight of 400, which is more
readable against the navy blue background.

Trophy text is now aligned at the top, solving the weirdness where the
title would look top-aligned and the clan and player lists aligned lower
than it.