Bugfix: apply streak retro-updates to clan points
Clan points for streaks were not retroactively updated along with player
games/points. Partly this was due to a lookup on
$player->get_score_by_key() where it should have been $clan (Streak.pm).
Correct that. Also change lookup key logic so both the clan and player
scoring entry lists are keyed with playername-n, where n is the player's
nth ascension. Finally, add a warning if Streak fails a lookup on a clan
game. Actually I'll add that warning if a player game lookup fails, too.
Partly the reason this was missed was silently ignoring failures that
shouldn't happen. OOPS. Player lookup fail warning will be added in a
--fixup. Behaviour is now correct in tests.

Reported-by: Luxidream
Signed-off-by: Joanna Doyle <jjadoyle@gmail.com>